\docType{data}
\name{dataMVpaperNoNA}
\alias{dataMVpaperNoNA}
\title{Clinical information on the 1404 patients of the multivariate paper}
\format{data frame with 1404 rows and 26 columns.}
\description{
  This data frame contains the clinical information
  relative to the 1404 patients analyzed in the
  multivariate paper. In this dataset the NAs are treated
  as a separate level when they appear in categorical
  variables.
}
\keyword{datasets}

